requirements.txt
asgiref==3.2.7
autopep8==1.5.2
beautifulsoup4==4.9.1
bs4==0.0.1
cached-property==1.5.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
commonmark==0.9.1
dictdiffer==0.8.1
Django==3.0.6
Flask==1.1.1
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
notion==0.0.25
oauthlib==3.1.0
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
python-slugify==4.0.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
